backend/requirements.txt
# FastAPI framework
fastapi>=0.104,<0.105
uvicorn[standard]>=0.24,<0.25

# Database
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
alembic>=1.12,<2.0
asyncpg>=0.29,<1.0

# Authentication
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.6,<1.0
itsdangerous>=2.1,<3.0

# Environment variables
python-dotenv>=1.0,<2.0

# File handling
aiofiles>=23.2,<24.0
Pillow>=10.0,<11.0
pdf2image>=1.16,<2.0
pytesseract>=0.3,<1.0

# Validation
pydantic[email]>=2.6,<3.0
email-validator>=2.1,<3.0

# CORS (modern usage in FastAPI)
# You can just use built-in CORS middleware, so no separate package needed

# Testing
pytest>=7.4,<8.0
httpx>=0.26,<1.0

# Utilities
uuid>=1.30,<2.0

# PDF Processing
PyPDF2
pdf2image
pytesseract

# OCR Support
tesseract-ocr  # System dependency
tesseract-ocr-amh  # Amharic language pack

# File Handling
python-magic
aiofiles



# Vector Database
pinecone-client

# Embeddings
openai
tiktoken

# RAG Framework
langchain
langchain-openai
langchain-pinecone

# Text Processing
sentence-transformers
nltk

# Utils
tenacity